Add the function get_dataset_semeval
This function can parse an XML document from the SemEval competition
2016 Task 5 Subtask 1. The reviews tested are the ones in English (from
the laptop and the restaurant domain).
The function can parse the document for the training part and the test
gold document.

The result is a dictionary which can be focused on the features or the
polarity of the sentences.