Use SemEval dataset from SubTask1 is possible
The config file has been modified to add the possibility of using the
new SemEval dataset.
The old_train.py will not be able to be used for the next version of
the algorithm, but it is kept to run future observations. The new
version of old_train.py is train.py.
It is now possible to run both training and prediction steps with the
SemEval dataset Subtask1.